add: 카테고리 api 호출 위치 변경 및 초기에 rendering 되는 조건 추가
