+++ frontend/client/src/components/Header/HeaderBottom/SearchContainer/AutoSearchList/AutoSearchList.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import { Link } from '@src/lib/CustomRouter/CustomRouter';
 import { AutoSearch } from '@src/types/Search';
